requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-seed==0.3.1
django-twilio==0.14.3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
drf-yasg==1.21.7
executing==1.2.0
Faker==19.1.0
idna==3.4
inflection==0.5.1
ipython==8.14.0
jedi==0.18.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
mysqlclient==2.2.0
packaging==23.1
parso==0.8.3
phonenumbers==8.13.16
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
toposort==1.10
traitlets==5.9.0
twilio==7.17.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
